APPLINK-3495

+++ src/components/application_manager/include/application_manager/event_engine/event_observer.h
@@ -47,7 +47,7 @@ class EventObserver
   friend class EventDispatcher;
 
   // Typedef for possible Observer ID's from mobile_apis functionID enum
-  typedef unsigned int ObserverID;
+  typedef unsigned long ObserverID;
 
   /*
    * @brief Constructor
