+++ src/components/application_manager/src/event_engine/event_observer.cc
@@ -39,7 +39,7 @@ namespace event_engine {
 EventObserver::EventObserver()
  : id_(0) {
   //Get unique id based on this
-  id_ = reinterpret_cast<unsigned int>(this);
+  id_ = reinterpret_cast<unsigned long>(this);
 }
 
 EventObserver::~EventObserver() {
